+++ src/main/resources/static/scripts/pages/interaction.js
@@ -278,7 +278,9 @@ function addInteractionRowButtonEvents(row) {
     const deleteButton = row.querySelector('.delete');
     const editButton = row.querySelector('.edit');
 
-    deleteButton.addEventListener('click', () => {
+    deleteButton.addEventListener('click', async () => {
+        if (!await showConfirmationModal("Você tem certeza que deseja excluir esta interação?")) return;
+
         deleteInteraction(row).catch(error => {
             console.error(error)
         });
